Refactored component folder structure Renamed Tickets component to TicketAlert
