+++ components/molecules/Countdown/Period.tsx
@@ -2,9 +2,9 @@ import * as React from 'react';
 import styled from '@emotion/styled';
 import { css } from 'emotion';
 import { DatespanUnit, DatespanPeriod } from './datespan';
-import { rhythm, Center } from '../typography';
+import { rhythm, Center } from '../../typography';
 import Arc from './Arc';
-import { breakpointMax } from '../variables';
+import { breakpointMax } from '../../variables';
 
 function unitString(unit: DatespanUnit, value: number) {
   const unitNames = {
