+++ components/molecules/ModalCard.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from '@emotion/styled';
-import { colors } from './variables';
-import { rhythm } from './typography';
+import { colors } from '../variables';
+import { rhythm } from '../typography';
 
 const Container = styled('section')({
   backgroundColor: `${colors.white}`,
