+++ components/organisms/TopNavigation/MobileNav.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react';
 import styled from '@emotion/styled';
 import { css } from 'emotion';
-import { colors, breakpointMin } from '../variables';
+import { colors, breakpointMin } from '../../variables';
 import { MenuItem } from '.';
-import { rhythm } from '../typography';
-import { NavButton } from '../Button';
+import { rhythm } from '../../typography';
+import { NavButton } from '../../atoms/Button';
 
 const transition = '200ms cubic-bezier(0.77, 0, 0.175, 1)';
 const iconWidth = 28.284;
