+++ pages/past-editions.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import styled from '@emotion/styled';
 import { DefaultPageProps } from './_app';
 import { Background } from '.';
-import TeamMembers from '../components/TeamMembers';
-import { SpeakersFlat } from '../components/Speakers';
+import TeamMembers from '../components/organisms/TeamMembers';
+import { SpeakersFlat } from '../components/organisms/Speakers';
 import Edition from '../types/Edition';
 import Speaker from '../types/Speaker';
 import TeamMember from '../types/TeamMember';
