+++ stories/AnimatedLogo.stories.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { storiesOf } from '@storybook/react';
-import AnimatedLogo from '../components/AnimatedLogo';
+import AnimatedLogo from '../components/atoms/AnimatedLogo';
 
 storiesOf('AnimatedLogo', module)
   .add('default view', () => (
