+++ stories/Button.stories.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
-import { Button, NavButton } from '../components/Button';
+import { Button, NavButton } from '../components/atoms/Button';
 
 storiesOf('Button', module)
   .add('default view', () => (
