+++ stories/LazyImage.stories.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { storiesOf } from '@storybook/react';
-import LazyImage from '../components/LazyImage';
+import LazyImage from '../components/atoms/LazyImage';
 import { action } from '@storybook/addon-actions';
 
 storiesOf('LazyImage', module)
