+++ stories/Loader.stories.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { storiesOf } from '@storybook/react';
-import Loader from '../components/Loader';
+import Loader from '../components/molecules/Loader';
 
 class LoadingApp extends React.Component {
   state = {
