+++ stories/Spinner.stories.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { storiesOf } from '@storybook/react';
-import Spinner from '../components/Spinner';
+import Spinner from '../components/atoms/Spinner';
 
 storiesOf('Spinner', module)
   // set position to 'relative' to override 'absolute'
